Update createfiles.py
Added support for all kind of files

+++ createfiles.py
@@ -6,20 +6,27 @@ filetype = "txt"
 
 def start():
     name=input("What name should the files have? ")
-    filetype=input("What should the type of file be? (soon, currently only supporting txt) ")
+    filetype=input("What should the type of file be? ")
     try:
-        amount=int(input("How many files should be generated?"))
+        amount=int(input("How many files should be created? "))
     except:
         print("Please enter a valid number!")
         start()
-    check=input(f"Are you sure you want to create {amount} files with the name {name} of the type {filetype}? (y/n)")
+    check=input(f"""Are you sure you want to create {amount} files with the name "{name}" of with the filetype (ending) {filetype}? (y/n)""")
     if check == "y":
-        for counter in range(amount):
-            fp = open(f'{name}{counter}.{filetype}', 'w')
-            fp.write(f'File number {counter}')
-            fp.close()
-            print(f"File number {counter} created")
-            counter+=1
+        if filetype == "txt":
+            for counter in range(amount):
+                fp = open(f'{name}{counter}.{filetype}', 'w')
+                fp.write(f'File number {counter}')
+                fp.close()
+                print(f"File number {counter} created")
+                counter+=1
+        else:
+            for counter in range(amount):
+                fp = open(f'{name}{counter}.{filetype}', 'w')
+                fp.close()
+                print(f"File number {counter} created")
+                counter+=1
     elif check == "n":
         print("Exiting...")
         sys.exit()
